Updated text of Google_Trends_notes.md
Added reference to Google Correlate.  (I played with it a little but didn't spend enough time to get my own time series to work. To pursue further, I would suggest downloading a csv of Google Trends data and copying the format in that file when pasting in our own stock return data.)