Minor text updates & added print commands at end
I just cleaned up a few small items in the text and added a print
command for the table display at the end.  (I did not install the google
trends reading package yet, so haven't run it properly yet.)  Either you
could re-run now it or I can install the package and run it later for it
to show the results again. -MB